requirements.txt
torch==1.7.1
torchvision==0.9.2
pycuda
tensorboard
pyyaml
tqdm
matplotlib
scipy
clean-fid
pandas